% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_registration.R
\name{reverseRegistration}
\alias{reverseRegistration}
\title{Reverse a registration}
\usage{
reverseRegistration(registration)
}
\arguments{
\item{An}{existing \link{Registration} object.}
}
\value{
The reversed registration object.
}
\description{
This function inverts the sense of an existing registration, swapping the
source and target images and inverting the associated transforms.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
