% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_registration.R
\docType{class}
\name{TransformSet-class}
\alias{TransformSet-class}
\alias{TransformSet}
\title{The TransformSet class}
\description{
This class represents a set of transformations relating to a single 2D or 3D
registration. They may encapsulate linear (affine) and/or nonlinear mappings
between the corresponding source and target spaces, and are compatible with
the \code{RNiftyReg} package.
}
\section{Fields}{

\describe{
\item{\code{objects}}{A list of transform objects, which are stored internally in
the format returned by the \code{\link[RNiftyReg:saveTransform]{RNiftyReg::saveTransform()}} function.}
}}


