% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_registration.R
\docType{data}
\name{TransformTypes}
\alias{TransformTypes}
\title{Valid transform types}
\format{
An object of class \code{character} of length 3.
}
\usage{
TransformTypes
}
\description{
Valid transform types
}
\keyword{datasets}
