% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_registration.R
\name{readRegistration}
\alias{readRegistration}
\title{Read a registration from file}
\usage{
readRegistration(path, validate = TRUE)
}
\arguments{
\item{path}{A string giving the path to the file or directory.}

\item{validate}{Boolean value. If \code{TRUE}, the default, a deserialised
object is checked to make sure it is of class \link{Registration}.}
}
\value{
A registration object.
}
\description{
Read a complete registration from file, including all associated transforms.
This function handles flat file and directory (.xfmb) formats used by all
recent versions of TractoR.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
